requirements.txt
pydantic
python-dotenv
azure-ai-projects
azure-ai-agents
azure-identity
azure-identity-broker
azure-cosmos
pymongo
requests
azure-ai-evaluation